+++ src/min3d/objectPrimitives/Rectangle.java
@@ -30,7 +30,7 @@ public class Rectangle extends Object3dContainer
 			for (col = 0; col <= $segsW; col++)
 			{
 				this.vertices().addVertex(
-					(float)col*w - width5, (float)row*h - height5,0f,
+					col*w - width5, row*h - height5,0f,
 					(float)col/(float)$segsW, 1 - (float)row/(float)$segsH,
 					0,0,1f,
 					color.r, color.g, color.b, color.a
